third_party/WebKit/LayoutTests/accessibility/render-counter-text-expected.txt

This tests that the text for LayoutCounter is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS line1.childAtIndex(0).name is '1'
PASS line1.childAtIndex(1).name is ':'
PASS line1.childAtIndex(2).name is 'line1'
PASS line2.childAtIndex(0).name is '2'
PASS line2.childAtIndex(1).name is ':'
PASS line2.childAtIndex(2).name is 'line2'
PASS successfullyParsed is true

TEST COMPLETE
